% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootlog.R
\name{bootlog}
\alias{bootlog}
\title{Simluates networks from a lolog model and fits the model to these networks}
\usage{
bootlog(lolog, seed = 1, fits = 10, cores = 2, MoreArgs = NULL,
  keep_nets = F)
}
\description{
Simluates networks from a lolog model and fits the model to these networks
}
\examples{
bootlog()
}
