% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootlog_plot.R
\name{bootlog_plot}
\alias{bootlog_plot}
\title{Plots the bootstrapped SEs in a nice way}
\usage{
bootlog_plot(results, lolog, facet = TRUE)
}
\description{
Plots the bootstrapped SEs in a nice way
}
\examples{
bootlog_plot()
}
