% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootlog_process.R
\name{bootlog_process}
\alias{bootlog_process}
\title{Does bootstrapped SEs for LOLOG in a nice way}
\usage{
bootlog_process(lolog, seed = 1, fits = seq(10, 100, 10), cores = 2,
  MoreArgs = NULL)
}
\description{
Does bootstrapped SEs for LOLOG in a nice way
}
\examples{
bootlog_process()
}
